User can checkout using Stripe and has redirection
Added a PaymentSession (which was previously Order) which contains details of the current payment by user. Using this, we redirect user after stripe payment to /status in Node so we can update the current payment status of the user. Also, deleted the cart if payment is success else let it be the same.